Add shared websocket library to all dependents
