--- apps/websocket/external
@@ -1,1 +1,0 @@
-./internal
