--- apps/websocket/internal/valid-topic/index.js
@@ -1,11 +1,0 @@
-const topicRegexp = /^([a-z]+)\/(command|event)\/([a-z]+)$/;
-
-const validTopic = (topic) => {
-  if (topicRegexp.test(topic)) {
-    return true;
-  }
-
-  throw new Error(`${topic} is not a valid topic`);
-};
-
-export default validTopic;
