+++ apps/youtube/external/index.html
@@ -7,6 +7,7 @@
   <title>YouTube Player</title>
 
   <link rel="stylesheet" href="styles.css">
+  <script src="/websocket"></script>
 </head>
 
 <body>
@@ -29,7 +30,7 @@
     <view class="loading">Loading</view>
   </root>
 
-  <script type="module" src="app.js"></script>
+  <script src="app.js"></script>
 </body>
 
 </html>
