+++ apps/youtube/internal/index.html
@@ -3,6 +3,7 @@
 
 <head>
   <meta charset="UTF-8">
+  <script src="/websocket"></script>
   <title>YouTube Player</title>
 </head>
 
@@ -11,7 +12,7 @@
   <audio class="media"></audio>
   <script src="https://www.youtube.com/iframe_api"></script>
   <script src="youtube.js"></script>
-  <script src="internal.js" type="module"></script>
+  <script src="internal.js"></script>
 </body>
 
 </html>
