+++ services/downloader/lib/io/ws/index.js
@@ -1,38 +1,25 @@
 const { URL } = require('url');
-const WebSocket = require('ws');
+const createWebsocket = require('websocket').default;
 
 const logger = require('../../logger')('ws');
 
-const subscribedTopic = 'mediaRequest';
-const publishTopic = 'mediaAvailable';
+const subscribedTopic = 'downloader/command/request';
+const publishTopic = 'downloader/event/available';
 
 const webSocket = ({ host, downloader }) => {
-  const wsPath = `ws://${host.hostname}:8000`;
-  const ws = new WebSocket(wsPath);
+  const ws = createWebsocket();
   const callback = sendMessage(ws);
   const handler = handleMessage({ host, downloader, callback });
 
-  ws.on('open', () => logger.info(`Listening to web socket ${wsPath}`));
-
-  ws.addEventListener('message', function(evt) {
-    handler(
-      JSON.parse(evt.data)
-    );
-  });
+  ws.subscribe(subscribedTopic, handler);
 };
 
 const sendMessage = ws => (topic, payload) => (
-  ws.send(
-    JSON.stringify({ topic, payload })
-  )
+  ws.publish({ topic, payload })
 );
 
-const handleMessage = ({ host, downloader, callback }) => async ({ topic, payload }) => {
+const handleMessage = ({ host, downloader, callback }) => async ({ payload }) => {
   const { url: sourceUrl } = payload;
-
-  if (topic != subscribedTopic || typeof sourceUrl == 'undefined') {
-    return;
-  }
 
   const videoPath = await downloader(sourceUrl);
   const url = new URL(videoPath, host);
