+++ services/downloader/package.json
@@ -14,7 +14,7 @@
     "ip": "^1.1.5",
     "serve-index": "^1.9.1",
     "winston": "^3.0.0-rc",
-    "ws": "^4.0.0"
+    "websocket": "../../shared/websocket"
   },
   "devDependencies": {
     "prettier": "1.9.2"
