+++ services/manager/lib/io/http/index.js
@@ -33,12 +33,20 @@ const startServer = (server, port, name) => {
   ));
 };
 
-const mountExternal = (apps, port, index) => {
+const mountWebsocket = server => {
+  const path = pathLib.dirname(require.resolve('websocket'));
+
+  mountApp({ server, name: 'websocket', path });
+};
+
+const mountExternal = (apps, port) => {
   const server = express();
 
   apps.map(({ name, path }) => {
     mountApp({ server, name, path: pathLib.join(path, 'external') })
   });
+
+  mountWebsocket(server);
 
   startServer(server, port, 'Public');
 };
@@ -51,6 +59,8 @@ const mountInternal = (apps, port, publicPath) => {
   apps.map(({ name, path }) => (
     mountApp({ server, name, path: pathLib.join(path, 'internal') })
   ));
+
+  mountWebsocket(server);
 
   server.use(
     express.static(publicPath)
