+++ services/manager/package.json
@@ -13,6 +13,7 @@
     "chalk": "^2.3.0",
     "express": "^4.14.0",
     "ip": "^1.1.5",
-    "ws": "^1.1.1"
+    "ws": "^1.1.1",
+    "websocket": "../../shared/websocket"
   }
 }
