+++ services/rfid/index.js
@@ -1,4 +1,3 @@
-const ip = require('ip');
 const { URL } = require('url');
 
 const ws = require('./lib/ws');
@@ -6,9 +5,8 @@ const rfid = require('./lib/rfid');
 const notify = require('./lib/notify');
 
 const port = process.env.PORT || 5002;
-const host = new URL(`http://${ip.address()}:${port}`);
 
-const send = ws({ host }).send;
+const send = ws().send;
 const { error, presented, removed } = notify(send);
 
 // Start polling for cards
