+++ services/rfid/package.json
@@ -14,6 +14,6 @@
   },
   "dependencies": {
     "ip": "^1.1.5",
-    "ws": "^4.0.0"
+    "websocket": "../../shared/websocket"
   }
 }
